店铺营业状态设置业务代码开发

+++ sky-pojo/src/main/java/com/sky/dto/DishPageQueryDTO.java
@@ -13,8 +13,8 @@ public class DishPageQueryDTO implements Serializable {
 
     private String name;
 
-    //分类id
-    private Integer categoryId;
+    //分类id 2024/10/8 原先int->Long 解决了菜品分类 根据菜品分类id查询菜品信息前端页面不显示的情况
+    private Long categoryId;
 
     //状态 0表示禁用 1表示启用
     private Integer status;
